executor/requirements.txt
anthropic>=0.40.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
httpx>=0.26.0
